+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable wandb callback. ([#494](https://github.com/jina-ai/finetuner/pull/494))
 
+- Support log streaming in finetuner client. ([#504](https://github.com/jina-ai/finetuner/pull/504))
+
 ### Removed
 
 ### Changed